documentanalyzer/requirements.txt

fastapi==0.110.2
uvicorn[standard]==0.29.0
openai==1.25.0
python-dotenv==1.0.1
pytest==8.1.1

